# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses/add": {
      "post": {
        "description": "Creates a new course in the system.",
        "operationId": "CoursesController_createCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "409": {
            "description": "Course with this code already exists"
          }
        },
        "summary": "Create Course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/update": {
      "patch": {
        "description": "Updates details of an existing course in the system.",
        "operationId": "CoursesController_updateCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "409": {
            "description": "Course with this code or title already exists"
          }
        },
        "summary": "Update Course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/current": {
      "get": {
        "description": "Allows admins to view any student's courses for the current semester.",
        "operationId": "CoursesController_getStudentCurrentCoursesForAdmin",
        "parameters": [
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current Courses fetched successfully"
          },
          "400": {
            "description": "studentId is required for administrators"
          },
          "403": {
            "description": "You are not authorized to access"
          }
        },
        "summary": "Get Student Current Semester Courses",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses": {
      "get": {
        "description": "Fetch a list of courses. Filter by code, name, or status.",
        "operationId": "CoursesController_getCourses",
        "parameters": [
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Filter by course code",
            "schema": {
              "example": "CSE101",
              "type": "string"
            }
          },
          {
            "name": "title",
            "required": false,
            "in": "query",
            "description": "Filter by course title",
            "schema": {
              "example": "Introduction to Computer Science",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by Status (active - true/false)",
            "schema": {
              "example": "true",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter by course status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses fetched successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          }
        },
        "summary": "Get Filtered Courses",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/student/all": {
      "get": {
        "description": "Retrieve history of all courses the authenticated student is or has been registered in.",
        "operationId": "CoursesController_getAllStudentCourses",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "semester",
            "required": false,
            "in": "query",
            "description": "Filter by specific Semester ID",
            "schema": {
              "example": "20000008",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses fetched successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          }
        },
        "summary": "Get Student All Registered Courses",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/{courseId}/enroll": {
      "post": {
        "description": "Enrolls the authenticated student in the specified course ID for the current semester.",
        "operationId": "CoursesController_enrollStudent",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "description": "The unique course identifier",
            "schema": {
              "example": "50000004",
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Student enrolled in course successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "Already enrolled in this course"
          }
        },
        "summary": "Enroll Student in Course",
        "tags": [
          "Courses"
        ]
      }
    },
    "/courses/withdraw/{studentId}/{courseId}": {
      "delete": {
        "description": "Withdraws the authenticated student from the specified course for the current semester.",
        "operationId": "CoursesController_withdrawStudent",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "description": "The unique course identifier",
            "schema": {
              "example": "50000004",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Student withdrawn from course successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": "You are not registered for this course"
          }
        },
        "summary": "Withdraw Student from Course",
        "tags": [
          "Courses"
        ]
      }
    },
...
  },
  "info": {
    "title": "APIs Documentation for EasyLearn Web App",
    "description": "APIs description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Course Code",
            "example": "PHY101"
          },
          "title": {
            "type": "string",
            "description": "Course Title",
            "example": "Introduction to Physics"
          },
          "credit": {
            "type": "number",
            "description": "Course Credits",
            "example": 3
          },
          "active": {
            "type": "boolean",
            "description": "Status (active)",
            "example": true
          }
        },
        "required": [
          "code",
          "title",
          "credit"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique course identifier",
            "example": "50000004"
          },
          "code": {
            "type": "string",
            "description": "Course Code",
            "example": "PHY101"
          },
          "title": {
            "type": "string",
            "description": "Course Title",
            "example": "Introduction to Physics"
          },
          "credit": {
            "type": "number",
            "description": "Course Credits",
            "example": 3
          },
          "active": {
            "type": "boolean",
            "description": "Status (active)",
            "example": true
          }
        },
        "required": [
          "id",
          "code",
          "title",
          "credit",
          "active"
        ]
      },
...
    }
  }
}